requirements.txt
certifi==2019.11.28
chardet==3.0.4
dnspython==1.16.0
entrypoints==0.3
flake8==3.7.9
idna==2.8
mccabe==0.6.1
mysql-connector-python==8.0.19
protobuf==3.6.1
pycodestyle==2.5.0
pyflakes==2.1.1
requests==2.22.0
six==1.14.0
urllib3==1.25.7
